Refactored data registration data sent to server, Removed dummy UserDetails page and integrated the data in Dashboard
