+++ app/src/main/java/me/twodee/friendlyneighbor/DashboardActivity.java
@@ -1,13 +1,32 @@
 package me.twodee.friendlyneighbor;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.graphics.Color;
+import android.net.Uri;
 import android.os.Bundle;
+import android.view.View;
 import android.view.WindowManager;
+import android.widget.Button;
+import android.widget.TextView;
+import android.widget.Toast;
+
+import com.google.android.gms.auth.api.signin.GoogleSignIn;
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
+import com.google.android.gms.auth.api.signin.GoogleSignInClient;
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
 
 public class DashboardActivity extends AppCompatActivity {
+
+    GoogleSignInClient mGoogleSignInClient;
+    Button sign_out;
+    TextView nameTV, emailTV;
+
+    String personName, personEmail;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -16,6 +35,57 @@ public class DashboardActivity extends AppCompatActivity {
                 WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS);
         getWindow().setStatusBarColor(Color.TRANSPARENT);
         setContentView(R.layout.activity_dashboard);
+
+        sign_out = findViewById(R.id.sign_out_button);
+        nameTV = findViewById(R.id.name);
+        emailTV = findViewById(R.id.email);
+
+        // Configure sign-in to request the user's ID, email address, and basic
+        // profile. ID and basic profile are included in DEFAULT_SIGN_IN.
+        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                .requestIdToken(getString(R.string.server_client_id))
+                .requestEmail()
+                .build();
+
+        // Build a GoogleSignInClient with the options specified by gso.
+        mGoogleSignInClient = GoogleSignIn.getClient(this, gso);
+
+        GoogleSignInAccount acct = GoogleSignIn.getLastSignedInAccount(DashboardActivity.this);
+        if (acct != null) {
+            personName = acct.getDisplayName();
+//            String personGivenName = acct.getGivenName();
+//            String personFamilyName = acct.getFamilyName();
+            personEmail = acct.getEmail();
+//            String personId = acct.getId();
+//            Uri personPhoto = acct.getPhotoUrl();
+
+            nameTV.setText("Welcome, " +personName);
+            emailTV.setText("Email: "+personEmail);
+//            idTV.setText("ID: "+personId);
+            //Glide.with(this).load(personPhoto).into(photoIV);
+
+        }
+
+
+
+        sign_out.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                signOut();
+            }
+        });
+    }
+
+    private void signOut() {
+        mGoogleSignInClient.signOut()
+                .addOnCompleteListener(this, new OnCompleteListener<Void>() {
+                    @Override
+                    public void onComplete(@NonNull Task<Void> task) {
+                        Toast.makeText(DashboardActivity.this,"Successfully Signed Out !!!",Toast.LENGTH_SHORT).show();
+                        startActivity(new Intent(DashboardActivity.this, MainActivity.class));
+                        finish();
+                    }
+                });
     }
 
     @Override
