+++ app/src/main/java/me/twodee/friendlyneighbor/RegistrationActivity.java
@@ -61,6 +61,7 @@ public class RegistrationActivity extends AppCompatActivity {
 
         //Validations
         awesomeValidation.addValidation(this, R.id.addressLine1, RegexTemplate.NOT_EMPTY, R.string.invalid_address_1);
+        awesomeValidation.addValidation(this, R.id.addressLine2, RegexTemplate.NOT_EMPTY, R.string.invalid_address_2);
         awesomeValidation.addValidation(this, R.id.city, RegexTemplate.NOT_EMPTY, R.string.invalid_city);
         awesomeValidation.addValidation(this, R.id.state, RegexTemplate.NOT_EMPTY, R.string.invalid_state);
         awesomeValidation.addValidation(this, R.id.country, RegexTemplate.NOT_EMPTY, R.string.invalid_country);
@@ -96,8 +97,8 @@ public class RegistrationActivity extends AppCompatActivity {
         try {
             //input your API parameters
             object.put("id", "5ebc27d7e6fe7a77013ecd2a");  // hardcoded for the time being
-            object.put("address1", addr1);
-            object.put("address2", addr2);
+            object.put("addressLineOne", addr1);
+            object.put("addressLineTwo", addr2);
             object.put("city", cty);
             object.put("state", st);
             object.put("country", cntry);
