--- app/src/main/java/me/twodee/friendlyneighbor/SignInUserDetails.java
@@ -1,115 +1,0 @@
-package me.twodee.friendlyneighbor;
-
-import android.content.Intent;
-import android.net.Uri;
-//import android.support.annotation.NonNull;
-//import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.ImageView;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-
-//import com.bumptech.glide.Glide;
-import com.android.volley.AuthFailureError;
-import com.android.volley.NetworkError;
-import com.android.volley.NoConnectionError;
-import com.android.volley.ParseError;
-import com.android.volley.Request;
-import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.ServerError;
-import com.android.volley.TimeoutError;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.StringRequest;
-import com.android.volley.toolbox.Volley;
-import com.google.android.gms.auth.api.signin.GoogleSignIn;
-import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
-import com.google.android.gms.auth.api.signin.GoogleSignInClient;
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-
-import java.net.URI;
-import java.util.HashMap;
-import java.util.Map;
-
-public class SignInUserDetails extends AppCompatActivity {
-
-    GoogleSignInClient mGoogleSignInClient;
-    Button sign_out;
-    TextView nameTV;
-    TextView emailTV;
-    TextView idTV;
-
-    String personName;
-    String personGivenName;
-    String personFamilyName;
-    String personEmail;
-    String personId;
-    Uri personPhoto;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_sign_in_user_details);
-
-        sign_out = findViewById(R.id.log_out);
-        nameTV = findViewById(R.id.name);
-        emailTV = findViewById(R.id.email);
-        idTV = findViewById(R.id.id);
-
-        // Configure sign-in to request the user's ID, email address, and basic
-        // profile. ID and basic profile are included in DEFAULT_SIGN_IN.
-        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestIdToken(getString(R.string.server_client_id))
-                .requestEmail()
-                .build();
-
-        // Build a GoogleSignInClient with the options specified by gso.
-        mGoogleSignInClient = GoogleSignIn.getClient(this, gso);
-
-        GoogleSignInAccount acct = GoogleSignIn.getLastSignedInAccount(SignInUserDetails.this);
-        if (acct != null) {
-            String personName = acct.getDisplayName();
-            String personGivenName = acct.getGivenName();
-            String personFamilyName = acct.getFamilyName();
-            String personEmail = acct.getEmail();
-            String personId = acct.getId();
-            Uri personPhoto = acct.getPhotoUrl();
-
-            nameTV.setText("Welcome, " +personName);
-            emailTV.setText("Email: "+personEmail);
-            idTV.setText("ID: "+personId);
-            //Glide.with(this).load(personPhoto).into(photoIV);
-
-            }
-
-
-
-        sign_out.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                signOut();
-            }
-        });
-    }
-
-
-    private void signOut() {
-        mGoogleSignInClient.signOut()
-                .addOnCompleteListener(this, new OnCompleteListener<Void>() {
-                    @Override
-                    public void onComplete(@NonNull Task<Void> task) {
-                        Toast.makeText(SignInUserDetails.this,"Successfully Signed Out !!!",Toast.LENGTH_SHORT).show();
-                        startActivity(new Intent(SignInUserDetails.this, MainActivity.class));
-                        finish();
-                    }
-                });
-    }
-}
