+++ app/src/main/res/layout/activity_dashboard.xml
@@ -59,12 +59,14 @@
                 android:layout_height="wrap_content">
 
                 <TextView
+                    android:id="@+id/name"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:textColor="@android:color/white"
                     android:textSize="24dp"
-                    android:text="Akhil Surendran"
+                    android:text="Akhil Surendran (default)"
                     android:fontFamily="@font/product_sans_bold"/>
+
 
                 <RelativeLayout
                     android:layout_width="match_parent"
@@ -78,8 +80,32 @@
 
                 </RelativeLayout>
 
+
             </LinearLayout>
 
+            <TextView
+                android:id="@+id/email"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textColor="@android:color/white"
+                android:textSize="12dp"
+                android:text="email (default)"
+                android:fontFamily="@font/product_sans_bold"/>
+
+            <RelativeLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="30dp">
+
+                <Button
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="Sign Out"
+                    android:id="@+id/sign_out_button"
+                    android:layout_centerInParent="true"/>
+
+
+            </RelativeLayout>
 
         </LinearLayout>
